refactor: reorganize frontend component structure
- Remove duplicate components (Topbar, PageHeader, AdminSidebar, ClientSidebar, UserGreeting, ClientDashboard)
- Remove unused components (LoadingModalDemo, AuthExample, ApiTest, NavAuditFirst, UserProfileDropdown, ClientRegisterModal, ClientLoginModal, LeadForm, CaseStudies, Faqs)
- Organize components into logical directories:
  * common/ - shared components (Header, Footer, SiteNav, LoadingModal)
  * features/ - feature components (HeroTulsaSEO, FeaturesSection, OurServicesStacked, PricingPage)
  * audit/ - audit components (AuditWizard)
  * portal/ - portal components (Topbar, PageHeader, AdminSidebar, ClientSidebar, UserGreeting, ClientDashboard)
  * ui/ - reusable UI components (button, card, input, sheet)
- Update all import statements to reflect new file locations
