+++ frontend/src/app/audit-wizard/page.tsx
@@ -1,4 +1,4 @@
-import AuditWizard from '@/components/AuditWizard';
+import AuditWizard from '@/components/audit/AuditWizard';
 
 export default function AuditWizardPage() {
   return <AuditWizard />;
