+++ frontend/src/app/contact/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import Header from '../../components/Header';
+import Header from '../../components/common/Header';
 
 export default function ContactPage() {
   return (
