+++ frontend/src/app/pricing/page.tsx
@@ -1,4 +1,4 @@
-import PricingPage from '@/components/PricingPage';
+import PricingPage from '@/components/features/PricingPage';
 
 export default function PricingRoute() {
   return <PricingPage />;
