+++ frontend/src/components/features/FeaturesSection.tsx
