+++ frontend/src/components/features/OurServicesStacked.tsx
