--- frontend/src/components/portal/UserProfileDropdown.tsx
@@ -1,224 +1,0 @@
-"use client";
-
-import { useState, useEffect, useRef } from "react";
-import { useRouter } from "next/navigation";
-import { fetchUserProfile, logoutUser, UserProfileData } from "../../services/userProfile";
-
-interface UserProfileDropdownProps {
-  className?: string;
-}
-
-export default function UserProfileDropdown({ className = "" }: UserProfileDropdownProps) {
-  const [isOpen, setIsOpen] = useState(false);
-  const [userProfile, setUserProfile] = useState<UserProfileData | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const dropdownRef = useRef<HTMLDivElement>(null);
-  const router = useRouter();
-
-  useEffect(() => {
-    // Load user profile data
-    const loadUserProfile = async () => {
-      try {
-        setLoading(true);
-        setError(null);
-        const profile = await fetchUserProfile();
-        setUserProfile(profile);
-      } catch (err) {
-        console.error('Failed to load user profile:', err);
-        setError(err instanceof Error ? err.message : 'Failed to load user data');
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    loadUserProfile();
-  }, []);
-
-  useEffect(() => {
-    // Close dropdown when clicking outside
-    const handleClickOutside = (event: MouseEvent) => {
-      if (dropdownRef.current && !dropdownRef.current.contains(event.target as Node)) {
-        setIsOpen(false);
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => document.removeEventListener('mousedown', handleClickOutside);
-  }, []);
-
-  const handleLogout = async () => {
-    try {
-      const result = await logoutUser();
-      if (result.success) {
-        // Redirect to login page
-        router.push(result.redirectUrl || '/login');
-      }
-    } catch (error) {
-      console.error('Logout failed:', error);
-      // Still redirect to login page even if logout fails
-      router.push('/login');
-    }
-  };
-
-  const handleNavigation = (path: string) => {
-    setIsOpen(false);
-    router.push(path);
-  };
-
-  // Get user initials for avatar
-  const getUserInitials = (): string => {
-    if (!userProfile) return "U";
-    
-    if (userProfile.user.firstName && userProfile.user.lastName) {
-      return `${userProfile.user.firstName[0]}${userProfile.user.lastName[0]}`.toUpperCase();
-    } else if (userProfile.user.name) {
-      const names = userProfile.user.name.split(' ');
-      if (names.length >= 2) {
-        return `${names[0][0]}${names[names.length - 1][0]}`.toUpperCase();
-      }
-      return names[0][0]?.toUpperCase() || "U";
-    }
-    
-    return userProfile.user.email[0]?.toUpperCase() || "U";
-  };
-
-  // Get display name
-  const getDisplayName = (): string => {
-    if (!userProfile) return "User";
-    
-    if (userProfile.user.name) return userProfile.user.name;
-    if (userProfile.user.firstName && userProfile.user.lastName) {
-      return `${userProfile.user.firstName} ${userProfile.user.lastName}`;
-    }
-    return userProfile.user.email;
-  };
-
-  if (loading) {
-    return (
-      <div className={`flex items-center space-x-3 ${className}`}>
-        <div className="w-8 h-8 bg-gray-300 rounded-full animate-pulse"></div>
-        <div className="w-24 h-4 bg-gray-300 rounded animate-pulse"></div>
-        <div className="w-4 h-4 bg-gray-300 rounded animate-pulse"></div>
-      </div>
-    );
-  }
-
-  if (error || !userProfile) {
-    return (
-      <div className={`flex items-center space-x-3 ${className}`}>
-        <div className="w-8 h-8 bg-red-100 rounded-full flex items-center justify-center">
-          <span className="text-red-600 text-sm font-medium">!</span>
-        </div>
-        <span className="text-gray-500 text-sm">Error loading profile</span>
-        <button
-          onClick={() => setIsOpen(!isOpen)}
-          className="text-gray-400 hover:text-gray-600 transition-colors"
-        >
-          <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-          </svg>
-        </button>
-      </div>
-    );
-  }
-
-  return (
-    <div className={`relative ${className}`} ref={dropdownRef}>
-      {/* User Profile Button */}
-      <button
-        onClick={() => setIsOpen(!isOpen)}
-        className="flex items-center space-x-3 p-2 rounded-lg hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors"
-      >
-        {/* User Avatar */}
-        <div className="w-8 h-8 bg-blue-600 rounded-full flex items-center justify-center text-white font-medium text-sm">
-          {getUserInitials()}
-        </div>
-        
-        {/* User Name */}
-        <span className="text-sm font-medium text-gray-900 dark:text-white">
-          {getDisplayName()}
-        </span>
-        
-        {/* Chevron Icon */}
-        <svg 
-          className={`w-4 h-4 text-gray-400 transition-transform ${isOpen ? 'rotate-180' : ''}`} 
-          fill="none" 
-          stroke="currentColor" 
-          viewBox="0 0 24 24"
-        >
-          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-        </svg>
-      </button>
-
-      {/* Dropdown Menu */}
-      {isOpen && (
-        <div className="absolute right-0 mt-2 w-64 bg-white dark:bg-gray-800 rounded-lg shadow-lg border border-gray-200 dark:border-gray-700 z-50">
-          {/* User Info Section */}
-          <div className="p-4 border-b border-gray-200 dark:border-gray-700">
-            <div className="flex items-center space-x-3">
-              <div className="w-10 h-10 bg-blue-600 rounded-full flex items-center justify-center text-white font-medium">
-                {getUserInitials()}
-              </div>
-              <div>
-                <p className="text-sm font-medium text-gray-900 dark:text-white">
-                  {getDisplayName()}
-                </p>
-                <p className="text-xs text-gray-500 dark:text-gray-400">
-                  {userProfile.user.email}
-                </p>
-              </div>
-            </div>
-          </div>
-
-          {/* Navigation Links */}
-          <div className="py-2">
-            <button
-              onClick={() => handleNavigation('/profile')}
-              className="w-full flex items-center space-x-3 px-4 py-2 text-sm text-gray-700 dark:text-gray-300 hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors"
-            >
-              <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
-              </svg>
-              <span>Profile</span>
-            </button>
-
-            <button
-              onClick={() => handleNavigation('/notifications')}
-              className="w-full flex items-center space-x-3 px-4 py-2 text-sm text-gray-700 dark:text-gray-300 hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors"
-            >
-              <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 17h5l-5 5v-5zM10.5 3.75a6 6 0 00-6 6v3.75l-.75.75V19.5h13.5v-5.25l-.75-.75V9.75a6 6 0 00-6-6z" />
-              </svg>
-              <span>Notifications</span>
-            </button>
-
-            <button
-              onClick={() => handleNavigation('/settings')}
-              className="w-full flex items-center space-x-3 px-4 py-2 text-sm text-gray-700 dark:text-gray-300 hover:bg-gray-100 dark:hover:bg-gray-700 transition-colors"
-            >
-              <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z" />
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
-              </svg>
-              <span>Settings</span>
-            </button>
-          </div>
-
-          {/* Sign Out Button */}
-          <div className="p-2 border-t border-gray-200 dark:border-gray-700">
-            <button
-              onClick={handleLogout}
-              className="w-full flex items-center justify-between px-4 py-2 text-sm text-red-600 dark:text-red-400 hover:bg-red-50 dark:hover:bg-red-900/20 rounded-md transition-colors"
-            >
-              <span>Sign out</span>
-              <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1" />
-              </svg>
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
