--- frontend/src/components/portal/dashboard/ClientDashboard.tsx
@@ -1,3 +1,0 @@
-export { default } from "../../portal/ClientDashboard";
-
-
