--- frontend/src/components/portal/layout/PageHeader.tsx
@@ -1,3 +1,0 @@
-export { default } from "../../portal/PageHeader";
-
-
