--- frontend/src/components/portal/user/UserGreeting.tsx
@@ -1,3 +1,0 @@
-export { default } from "../../portal/UserGreeting";
-
-
